fix(db): 统一 users 和 videos 表的 mid 字段类型为 BIGINT
将 `videos.mid` 类型从 INTEGER 改为 BIGINT，避免数据类型不一致导致外键约束失败

+++ init.sql
@@ -3,11 +3,10 @@ CREATE TABLE users (
     name TEXT NOT NULL
 );
 
-
 CREATE TABLE videos (
     aid BIGINT PRIMARY KEY,
     bvid TEXT NOT NULL UNIQUE,
-    mid INTEGER NOT NULL,
+    mid BIGINT NOT NULL,
     title TEXT NOT NULL,
     description TEXT,
     pic TEXT,
